+++ CHANGELOG.md
@@ -22,6 +22,8 @@ since `1.7.0`.
 - Added `argc` and `argv` to contexts so that they can optionaly support CLI
   commands.
 - A struct is used to pass paramaters for creating contexts.
+- The OpenGL 2 renderer can now use `SDL_HINT_RENDER_SCALE_QUALITY` to
+  determine the tileset upscaling filter.
 
 ### Fixed
 - `TCOD_console_is_key_pressed` should work even if libtcod events are ignored.